fixed "count may be uninitialised" compilation warning changed the return type of splitlines to size_t (instead of int)
git-svn-id: http://ezix.org/source/packages/lshw/development@714 811e2811-9fd9-0310-a116-b6e8ac943c8b

+++ osutils.h
@@ -14,7 +14,7 @@ bool samefile(const std::string & path1, const std::string & path2);
 std::string resolvesymlinks(const std::string & path);
 bool loadfile(const std::string & file, std::vector < std::string > &lines);
 
-int splitlines(const std::string & s,
+size_t splitlines(const std::string & s,
 		std::vector < std::string > &lines,
 		char separator = '\n');
 std::string get_string(const std::string & path, const std::string & def = "");
